LPS-152635 Dashboard-Tasks Refactor
