+++ modules/apps/site-initializer/site-initializer-raylife-ap/src/main/resources/site-initializer/fragments/src/raylife-ap-react-fragments/fragments/dashboard-tasks/index.css
@@ -1,47 +1,21 @@
-.dashboard-tasks-container .dashboard-tasks-content {
-	margin-top: 30px;
+.dashboard-tasks-container .dashboard-tasks-header {
+	margin-bottom: 30px;
 }
 
 .dashboard-tasks-container
 	.dashboard-tasks-header
 	.dashboard-tasks-title::after {
-	background-color: #4d85ff;
+	background-color: var(--primary);
 	content: '';
+	display: block;
 	height: 1px;
-	left: 28px;
-	min-width: 20px;
+	margin-top: 6px;
 	position: absolute;
-	top: 50px;
+	width: 19px;
 }
 
-.dashboard-tasks-container
-	.dashboard-tasks-section.border-bottom-dashed:not(:last-child) {
-	border-bottom: 1px dashed #d3d3d4;
-	padding-bottom: 10px;
-}
-
-.dashboard-tasks-container .dashboard-tasks-subsection .arrow-icon svg,
 .dashboard-tasks-container .dashboard-tasks-header .settings-icon svg {
 	cursor: pointer;
-	height: 24px !important;
-	width: 24px !important;
-}
-
-.dashboard-tasks-container .dashboard-tasks-subsection .bubble-element {
-	height: 24px !important;
-	width: 30px !important;
-}
-
-.dashboard-tasks-container .dashboard-tasks-subsection .bubble-element,
-.dashboard-tasks-container
-	.dashboard-tasks-content
-	.dashboard-tasks-section
-	.dashboard-tasks-subsection {
-	cursor: pointer;
-}
-
-.dashboard-tasks-container
-	.dashboard-tasks-section
-	.dashboard-tasks-section-header {
-	font-weight: 900 !important;
+	height: 20px !important;
+	width: 20px !important;
 }
